+++ src/utils/mips/wasm-worker.ts
@@ -1,0 +1,1 @@
+export {}
